* modules/proxy/ajp_msg.c (ajp_msg_dump): Cast apr_size_t to unsigned long to avoid the warning with %lx for 32-bit size_t too.
git-svn-id: https://svn.apache.org/repos/asf/httpd/httpd/trunk@105104 13f79535-47bb-0310-9956-ffa450edef68

+++ modules/proxy/ajp_msg.c
@@ -70,7 +70,7 @@ apr_status_t ajp_msg_dump(ajp_msg_t *msg, char *err)
 
         ap_log_error(APLOG_MARK, APLOG_DEBUG, 0, NULL,
                       "ajp_msg_dump(): %.4lx    %s",
-                      i, line);
+                      (unsigned long)i, line);
     }
     
     return APR_SUCCESS;
